+++ parser_library/src/workspaces/file_manager_impl.cpp
@@ -189,10 +189,11 @@ bool file_manager_impl::dir_exists(const utils::resource::resource_location& dir
     return utils::resource::dir_exists(dir_loc);
 }
 
-void file_manager_impl::put_virtual_file(unsigned long long id, std::string_view text)
+void file_manager_impl::put_virtual_file(
+    unsigned long long id, std::string_view text, utils::resource::resource_location related_workspace)
 {
     std::lock_guard guard(virtual_files_mutex);
-    m_virtual_files.try_emplace(id, text);
+    m_virtual_files.try_emplace(id, text, std::move(related_workspace));
 }
 
 void file_manager_impl::remove_virtual_file(unsigned long long id)
@@ -205,8 +206,16 @@ std::string file_manager_impl::get_virtual_file(unsigned long long id) const
 {
     std::lock_guard guard(virtual_files_mutex);
     if (auto it = m_virtual_files.find(id); it != m_virtual_files.end())
-        return it->second;
+        return it->second.text;
     return {};
+}
+
+utils::resource::resource_location file_manager_impl::get_virtual_file_workspace(unsigned long long id) const
+{
+    std::lock_guard guard(virtual_files_mutex);
+    if (auto it = m_virtual_files.find(id); it != m_virtual_files.end())
+        return it->second.related_workspace;
+    return utils::resource::resource_location();
 }
 
 } // namespace hlasm_plugin::parser_library::workspaces