+++ parser_library/src/workspaces/file_manager_vfm.h
@@ -17,6 +17,7 @@
 
 #include <string_view>
 
+#include "utils/resource_location.h"
 #include "virtual_file_monitor.h"
 
 namespace hlasm_plugin::parser_library::workspaces {
@@ -25,12 +26,14 @@ class file_manager;
 struct file_manager_vfm final : public virtual_file_monitor
 {
     file_manager& fm;
+    hlasm_plugin::utils::resource::resource_location related_workspace;
 
     // Inherited via virtual_file_monitor
     virtual_file_handle file_generated(std::string_view content) override;
 
-    explicit file_manager_vfm(file_manager& fm)
+    explicit file_manager_vfm(file_manager& fm, hlasm_plugin::utils::resource::resource_location related_workspace)
         : fm(fm)
+        , related_workspace(std::move(related_workspace))
     {}
 
 private: