+++ parser_library/src/workspaces/workspace_configuration.h
@@ -25,6 +25,7 @@
 #include <unordered_map>
 #include <utility>
 #include <vector>
+#include <version>
 
 #include "config/b4g_config.h"
 #include "config/pgm_conf.h"