+++ parser_library/test/debugging/debugger_test.cpp
@@ -952,7 +952,7 @@ TEST(debugger, ainsert)
     } };
 
     step_over_by(3, d, m, exp_frames, 12);
-    step_into(d, m, exp_frames, 2, "MACRO", resource_location("hlasm://0/AINSERT:1.hlasm"));
+    step_into(d, m, exp_frames, 2, "MACRO", resource_location("hlasm://0/AINSERT_1.hlasm"));
     exp_frame_vars.insert(exp_frame_vars.begin(), frame_vars_ignore_sys_vars({}, {}, {}));
 
     EXPECT_TRUE(check_step(d, exp_frames, exp_frame_vars));