+++ parser_library/test/processing/ainsert_test.cpp
@@ -126,7 +126,7 @@ A        DC C
 
     EXPECT_EQ(d.code, "D016");
     ASSERT_EQ(d.related.size(), 3);
-    EXPECT_EQ(d.related[0].location.uri, "AINSERT:1.hlasm");
+    EXPECT_EQ(d.related[0].location.uri, "AINSERT_1.hlasm");
     EXPECT_EQ(d.related[1].location.uri, "COPYBOOK");
 }